Clean up errors, utils folder
Created simple error handler middleware. Created a utility folder and added validateRegistration and jwtVerify to authUtils.js
