+++ server/app.js
@@ -6,7 +6,7 @@ const logger = require("morgan");
 
 const indexRouter = require("./routes/index");
 const pingRouter = require("./routes/ping");
-const authRoutes = require("./routes/auth").router;
+const authRoutes = require("./routes/auth");
 
 const app = express();
 
@@ -20,20 +20,9 @@ app.use("/", indexRouter);
 app.use("/ping", pingRouter);
 app.use(authRoutes);
 
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
+// Error handler
 app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get("env") === "development" ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.json({ error: err });
+  res.status(401).send({ success: false, error: err })
 });
 
 app.listen(4000, () => {
