+++ server/routes/index.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 
 // Middleware to verify users on protected routes
-const jwtVerify = require('./auth').jwtVerify;
+const { jwtVerify } = require('../utils/authUtils');
 
 router.get("/welcome", function(req, res, next) {
   res.status(200).send({ welcomeMessage: "Step 1 (completed)" });
